Move NotAuthorised component to authPages folder
Relocated the NotAuthorised.tsx file from the general pages directory to the authPages subdirectory. This change better organizes the project structure by categorizing pages related to authorization under the authPages directory.
