+++ frontend/src/App.tsx
@@ -5,7 +5,7 @@ import LoginPage from "./pages/authPages/LoginPage.tsx";
 import RegisterPage from "./pages/authPages/RegisterPage.tsx";
 import Home from "./pages/userPages/Home.tsx";
 import WelcomePage from "./pages/WelcomePage.tsx";
-import NotAuthorised from "./pages/Not-Authorised.tsx";
+import NotAuthorised from "./pages/authPages/Not-Authorised.tsx";
 import TransferForm from "./pages/userPages/TransferForm.tsx";
 import AccountTransactions from "./pages/userPages/AccountTransactions.tsx";
 import CurrencyConverterForm from "./pages/userPages/CurrencyConverterForm.tsx";
